tag_extractor/requirements.txt
flask-1.1.2
bs4-4.9.3
requests-2.25.1
konlpy-0.5.2